Decompose text game into separate screen handlers
